# internal/httpapi/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Job Tracker API
  version: 1.0.0
  description: API for tracking job applications and their status transitions.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/applications:
    post:
      summary: Create an application
      operationId: createApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
            example:
              companyName: Acme
              title: Backend Engineer
              source: LinkedIn
              jobUrl: https://example.com/jobs/backend-engineer
              location: Remote
              currentStatus: APPLIED
              notes: Follow up next week.
      responses:
        "201":
          description: Application created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      summary: List applications
      operationId: listApplications
      responses:
        "200":
          description: Applications returned.
          content:
            application/json:
              schema:
                type: object
                required:
                  - applications
                  - total
                properties:
                  applications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Application"
                  total:
                    type: integer
                    example: 1
  /api/v1/applications/{id}:
    get:
      summary: Get an application
      operationId: getApplication
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      responses:
        "200":
          description: Application returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Replace an application
      description: Updates the whole application while preserving server-managed fields such as id and createdAt.
      operationId: updateApplication
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
            example:
              companyName: Acme Updated
              title: Senior Backend Engineer
              source: Company site
              location: Remote
              currentStatus: TECHNICAL_INTERVIEW
              notes: Updated details.
      responses:
        "200":
          description: Application updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/applications/{id}/status-transitions:
    post:
      summary: Create a status transition
      operationId: createStatusTransition
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStatusTransitionRequest"
            example:
              toStatus: TECHNICAL_INTERVIEW
              note: First technical screen scheduled.
      responses:
        "201":
          description: Status transition created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusTransition"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      summary: List status transitions
      operationId: listStatusTransitions
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      responses:
        "200":
          description: Status transitions returned.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - total
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/StatusTransition"
                  total:
                    type: integer
                    example: 1
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    ApplicationID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 8ef78c64-1234-4f7e-89fb-0ae992af937d
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ApplicationStatus:
      type: string
      enum:
        - SAVED
        - APPLIED
        - RECRUITER_SCREEN
        - ONLINE_ASSESSMENT
        - TECHNICAL_INTERVIEW
        - FINAL_ROUND
        - OFFER
        - REJECTED
        - WITHDRAWN
        - NO_RESPONSE
    Salary:
      type: object
      properties:
        min:
          type: integer
          example: 90000
        max:
          type: integer
          example: 130000
        currency:
          type: string
          default: USD
          example: USD
    JobDescription:
      type: object
      required:
        - contentType
        - content
      properties:
        contentType:
          type: string
          example: text/plain
        content:
          type: string
          example: Build APIs and backend services.
    ApplicationInput:
      type: object
      required:
        - companyName
        - title
      properties:
        companyName:
          type: string
          example: Acme
        title:
          type: string
          example: Backend Engineer
        source:
          type: string
          example: LinkedIn
        jobUrl:
          type: string
          format: uri
          example: https://example.com/jobs/backend-engineer
        location:
          type: string
          example: Remote
        jobDescription:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/JobDescription"
        salary:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Salary"
        sponsorship:
          type: string
          example: Unknown
        currentStatus:
          $ref: "#/components/schemas/ApplicationStatus"
        applicationSubmittedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          example: Follow up next week.
    Application:
      allOf:
        - $ref: "#/components/schemas/ApplicationInput"
        - type: object
          required:
            - id
            - companyName
            - title
            - currentStatus
            - createdAt
            - updatedAt
          properties:
            id:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    CreateStatusTransitionRequest:
      type: object
      required:
        - toStatus
      properties:
        toStatus:
          $ref: "#/components/schemas/ApplicationStatus"
        note:
          type: string
          example: First technical screen scheduled.
    StatusTransition:
      type: object
      required:
        - id
        - applicationId
        - toStatus
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        applicationId:
          type: string
          format: uuid
        fromStatus:
          $ref: "#/components/schemas/ApplicationStatus"
        toStatus:
          $ref: "#/components/schemas/ApplicationStatus"
        note:
          type: string
          example: First technical screen scheduled.
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: application not found